Remove C bitwise operations from gate functions
The built-in C bitwise operations have been removed from
the logic gate functions, in favor of if-statement based
decision-making.

The purpose of this change is to dig deeper into the logic
of the basic gates, and to ensure that more complex gates,
such as the XOR gate, are implemented by chaining together basic gates,
rather than by using the built-in C operators.